+++ R/components.R
@@ -26,7 +26,8 @@
 #' @rdname components
 #' @export
 components.mdl_df <- function(object, ...){
-  object <- gather(object, ".model", ".fit", !!!syms(mable_vars(object)))
+  object <- tidyr::pivot_longer(object, all_of(mable_vars(object)),
+                             names_to = ".model", values_to = ".fit")
   kv <- key_vars(object)
   object <- transmute(as_tibble(object),
                       !!!syms(kv), !!sym(".model"),