+++ R/dable.R
@@ -7,7 +7,6 @@
 #' `autoplot()` method for displaying decompositions. Beyond this, a dable
 #' (`dcmp_ts`) behaves very similarly to a tsibble (`tbl_ts`).
 #'
-#' @inheritParams fable
 #' @param ... Arguments passed to [tsibble::tsibble()].
 #' @param response The name of the response variable column.
 #' @param method The name of the decomposition method.
@@ -96,7 +95,7 @@ tbl_sum.dcmp_ts <- function(x){
 
 #' @export
 rbind.dcmp_ts <- function(...){
-  .Deprecated("bind_rows()")
+  deprecate_warn("0.2.0", "rbind.fbl_ts()", "bind_rows()")
   dots <- dots_list(...)
   
   attrs <- combine_dcmp_attr(dots)