+++ R/dplyr-fable.R
@@ -24,3 +24,23 @@ dplyr_reconstruct.fbl_ts <- function(data, template) {
 
 #' @export
 dplyr_reconstruct.grouped_fbl <- dplyr_reconstruct.fbl_ts
+
+#' @export
+summarise.fbl_ts <- function(.data, ..., .groups = NULL) { 
+  dist_var <- distribution_var(.data)
+  dist_ptype <- vec_ptype(.data[[dist_var]])
+  resp_var <- response_vars(.data)
+  .data <- summarise(as_tsibble(.data), ..., .groups = .groups)
+  
+  # If the distribution is lost, return a tsibble
+  if(!(dist_var %in% names(.data))) {
+    if(!vec_is(.data[[dist_var]], dist_ptype)){
+      return(.data)
+    }
+  }
+  
+  build_fable(.data, response = resp_var, distribution = dist_var)
+}
+
+#' @export
+summarise.grouped_fbl <- summarise.fbl_ts