+++ R/fabletools-package.R
@@ -12,4 +12,6 @@ globalVariables(".")
 #' @importFrom dplyr dplyr_row_slice dplyr_col_modify dplyr_reconstruct
 #' @importFrom dplyr bind_rows bind_cols
 #' @importFrom tidyr nest unnest gather spread
+#' @importFrom tidyselect all_of
+#' @importFrom lifecycle deprecate_warn
 NULL