+++ R/features.R
@@ -28,13 +28,13 @@ features_impl <- function(.tbl, .var, features, ...){
       res <- transpose(map(key_dt[[".rows"]], function(i){
         out <- do.call(fn_safe, c(list(x[i]), dots[intersect(names(fmls), names(dots))]))
         if(is.null(names(out[["result"]]))) 
-          names(out[["result"]]) <- rep(".?", length(out[["result"]]))
+          names(out[["result"]]) <- paste0("..?", seq_along(out[["result"]]))
         out
       }))
       err <- compact(res[["error"]])
       tbl <- vctrs::vec_rbind(!!!res[["result"]])
       
-      names(tbl)[names(tbl) == ".?"] <- ""
+      names(tbl)[grepl("^\\.\\.?", names(tbl))] <- ""
       if(is.character(nm) && nzchar(nm)){
         names(tbl) <- sprintf("%s%s%s", nm, ifelse(nzchar(names(tbl)), "_", ""), names(tbl))
       }
@@ -73,7 +73,8 @@ features_impl <- function(.tbl, .var, features, ...){
   
   bind_cols(
     key_dt[-NCOL(key_dt)],
-    !!!out
+    !!!out,
+    .name_repair = "minimal"
   )
 }