+++ R/generate.R
@@ -28,12 +28,13 @@
 #' }
 #' @export
 generate.mdl_df <- function(x, new_data = NULL, h = NULL, times = 1, seed = NULL, ...){
-  kv <- c(key_vars(x), ".model")
   mdls <- mable_vars(x)
   if(!is.null(new_data)){
     new_data <- bind_new_data(x, new_data)[["new_data"]]
   }
-  x <- as_tibble(gather(x, ".model", ".sim", !!!syms(mdls)))
+  kv <- c(key_vars(x), ".model")
+  x <- tidyr::pivot_longer(as_tibble(x), all_of(mdls),
+                           names_to = ".model", values_to = ".sim")
   
   # Evaluate simulations
   x$.sim <- map2(x[[".sim"]],