+++ R/reconciliation.R
@@ -58,7 +58,6 @@ min_trace <- function(models, method = c("wls_var", "ols", "wls_struct", "mint_c
             method = match.arg(method), sparse = sparse)
 }
 
-#' @importFrom utils combn
 #' @export
 forecast.lst_mint_mdl <- function(object, key_data, 
                                   new_data = NULL, h = NULL,
@@ -67,13 +66,11 @@ forecast.lst_mint_mdl <- function(object, key_data,
   sparse <- object%@%"sparse"
   if(sparse){
     require_package("Matrix")
-    require_package("methods")
     as.matrix <- Matrix::as.matrix
     t <- Matrix::t
     diag <- function(x) if(is.vector(x)) Matrix::Diagonal(x = x) else Matrix::diag(x)
     solve <- Matrix::solve
     cov2cor <- Matrix::cov2cor
-    rowSums <- Matrix::rowSums
   } else {
     cov2cor <- stats::cov2cor
   }
@@ -85,11 +82,6 @@ forecast.lst_mint_mdl <- function(object, key_data,
   if(length(unique(map(fc, interval))) > 1){
     abort("Reconciliation of temporal hierarchies is not yet supported.")
   }
-  fc_dist <- map(fc, function(x) x[[distribution_var(x)]])
-  is_normal <- all(map_lgl(fc_dist, function(x) inherits(x[[1]], "dist_normal")))
-  
-  fc_mean <- as.matrix(invoke(cbind, map(fc_dist, mean)))
-  fc_var <- transpose_dbl(map(fc_dist, distributional::variance))
   
   # Compute weights (sample covariance)
   res <- map(object, function(x, ...) residuals(x, ...), type = "response")
@@ -141,9 +133,6 @@ forecast.lst_mint_mdl <- function(object, key_data,
   }
   
   # Reconciliation matrices
-  R1 <- cov2cor(W)
-  W_h <- map(fc_var, function(var) diag(sqrt(var))%*%R1%*%t(diag(sqrt(var))))
-  
   if(sparse){ 
     row_btm <- agg_data$leaf
     row_agg <- seq_len(nrow(key_data))[-row_btm]
@@ -170,44 +159,143 @@ forecast.lst_mint_mdl <- function(object, key_data,
     P <- solve(R%*%S)%*%R
   }
   
-  # Apply to forecasts
-  fc_mean <- as.matrix(S%*%P%*%t(fc_mean))
-  fc_mean <- split(fc_mean, row(fc_mean))
-  if(is_normal){
-    fc_var <- map(W_h, function(W) diag(S%*%P%*%W%*%t(P)%*%t(S)))
-    fc_dist <- map2(fc_mean, transpose_dbl(map(fc_var, sqrt)), distributional::dist_normal)
-  } else {
-    fc_dist <- map(fc_mean, distributional::dist_degenerate)
-  }
-  
-  # Update fables
-  map2(fc, fc_dist, function(fc, dist){
-    dimnames(dist) <- dimnames(fc[[distribution_var(fc)]])
-    fc[[distribution_var(fc)]] <- dist
-    point_fc <- compute_point_forecasts(dist, point_method)
-    fc[names(point_fc)] <- point_fc
-    fc
-  })
+  reconcile_fbl_list(fc, S, P, W, point_forecast = point_method)
 }
 
+#' Bottom up forecast reconciliation
+#' 
+#' \lifecycle{experimental}
+#' 
+#' Reconciles a hierarchy using the bottom up reconciliation method. The 
+#' response variable of the hierarchy must be aggregated using sums. The 
+#' forecasted time points must match for all series in the hierarchy.
+#' 
+#' @param models A column of models in a mable.
+#' 
+#' @seealso 
+#' [`reconcile()`], [`aggregate_key()`]
+#' @export
 bottom_up <- function(models){
   structure(models, class = c("lst_btmup_mdl", "lst_mdl", "list"))
 }
 
-#' @importFrom utils combn
 #' @export
 forecast.lst_btmup_mdl <- function(object, key_data, 
                                    point_forecast = list(.mean = mean), ...){
   # Keep only bottom layer
-  S <- build_smat_rows(key_data)
-  object <- object[rowSums(S) == 1]
+  agg_data <- build_key_data_smat(key_data)
+  
+  S <- matrix(0L, nrow = length(agg_data$agg), ncol = max(vec_c(!!!agg_data$agg)))
+  S[length(agg_data$agg)*(vec_c(!!!agg_data$agg)-1) + rep(seq_along(agg_data$agg), lengths(agg_data$agg))] <- 1L
+  
+  btm <- agg_data$leaf
+  object <- object[btm]
   
   point_method <- point_forecast
   point_forecast <- list()
+  
   # Get base forecasts
-  fc <- NextMethod()
+  fc <- vector("list", nrow(S))
+  fc[btm] <- NextMethod()
+  
+  # Add dummy forecasts to unused levels
+  fc[seq_along(fc)[-btm]] <- fc[btm[1]]
+  
+  P <- matrix(0L, nrow = ncol(S), ncol = nrow(S))
+  P[(btm-1L)*nrow(P) + seq_len(nrow(P))] <- 1L
+  
+  reconcile_fbl_list(fc, S, P, W = diag(nrow(S)),
+                     point_forecast = point_method)
+}
+
+
+#' Top down forecast reconciliation
+#' 
+#' \lifecycle{experimental}
+#' 
+#' Reconciles a hierarchy using the top down reconciliation method. The 
+#' response variable of the hierarchy must be aggregated using sums. The 
+#' forecasted time points must match for all series in the hierarchy.
+#' 
+#' @param models A column of models in a mable.
+#' @param method The reconciliation method to use.
+#' 
+#' @seealso 
+#' [`reconcile()`], [`aggregate_key()`]
+#' 
+#' @export
+top_down <- function(models, method = c("forecast_proportions", "average_proportions", "proportion_averages")){
+  structure(models, class = c("lst_topdwn_mdl", "lst_mdl", "list"),
+            method = match.arg(method))
+}
+
+#' @export
+forecast.lst_topdwn_mdl <- function(object, key_data, 
+                                   point_forecast = list(.mean = mean), ...){
+  method <- object%@%"method"
+  point_method <- point_forecast
+  point_forecast <- list()
+  
+  # TODO: Add check for grouped hierarchies
+  agg_data <- build_key_data_smat(key_data)
+  S <- matrix(0L, nrow = length(agg_data$agg), ncol = max(vec_c(!!!agg_data$agg)))
+  S[length(agg_data$agg)*(vec_c(!!!agg_data$agg)-1) + rep(seq_along(agg_data$agg), lengths(agg_data$agg))] <- 1L
+  
+  # Identify top and bottom level
+  top <- which.max(rowSums(S))
+  btm <- which(rowSums(S) == 1L)
+  
+  if(method == "forecast_proportions") {
+    abort("`method = 'forecast_proportions'` is not yet supported")
+    fc <- NextMethod()
+    fc_mean <- lapply(fc, function(x) mean(x[[distribution_var(x)]]))
+  } else {
+    # Compute dis-aggregation matrix
+    history <- lapply(object, function(x) response(x)[[".response"]])
+    top_y <- history[[top]]
+    btm_y <- history[btm]
+    if (method == "average_proportions") { 
+      prop <- map_dbl(btm_y, function(y) mean(y/top_y))
+    } else if (method == "proportion_averages") {
+      prop <- map_dbl(btm_y, mean) / mean(top_y)
+    } else {
+      abort("Unkown `top_down()` reconciliation `method`.")
+    }
+    
+    # Keep only top layer
+    object <- object[top]
+    
+    # Get base forecasts
+    fc <- vector("list", nrow(S))
+    fc[top] <- NextMethod()
+    
+    # Add dummy forecasts to unused levels
+    fc[seq_along(fc)[-top]] <- fc[top]
+  }
+  
+  P <- matrix(0L, nrow = ncol(S), ncol = nrow(S))
+  P[,top] <- prop
+  
+  reconcile_fbl_list(fc, S, P, W = diag(nrow(S)),
+                     point_forecast = point_method)
+}
+
+reconcile_fbl_list <- function(fc, S, P, W, point_forecast, SP = NULL) {
   if(length(unique(map(fc, interval))) > 1){
     abort("Reconciliation of temporal hierarchies is not yet supported.")
+  }
+  if(!inherits(S, "matrix")) {
+    # Use sparse functions
+    require_package("Matrix")
+    as.matrix <- Matrix::as.matrix
+    t <- Matrix::t
+    diag <- function(x) if(is.vector(x)) Matrix::Diagonal(x = x) else Matrix::diag(x)
+    cov2cor <- Matrix::cov2cor
+  } else {
+    cov2cor <- stats::cov2cor
+  }
+  if(is.null(SP)) {
+    SP <- S%*%P
   }
   
   fc_dist <- map(fc, function(x) x[[distribution_var(x)]])
@@ -217,26 +305,29 @@ forecast.lst_btmup_mdl <- function(object, key_data,
   fc_var <- transpose_dbl(map(fc_dist, distributional::variance))
   
   # Apply to forecasts
-  fc_mean <- as.matrix(S%*%t(fc_mean))
+  fc_mean <- as.matrix(SP%*%t(fc_mean))
   fc_mean <- split(fc_mean, row(fc_mean))
   if(is_normal){
-    fc_var <- map(fc_var, function(W) diag(S%*%diag(W)%*%t(S)))
+    R1 <- cov2cor(W)
+    W_h <- map(fc_var, function(var) diag(sqrt(var))%*%R1%*%t(diag(sqrt(var))))
+    fc_var <- map(W_h, function(W) diag(SP%*%W%*%t(SP)))
     fc_dist <- map2(fc_mean, transpose_dbl(map(fc_var, sqrt)), distributional::dist_normal)
   } else {
     fc_dist <- map(fc_mean, distributional::dist_degenerate)
   }
   
   # Update fables
-  pmap(list(rep_along(fc_mean, fc[1]), fc_mean, fc_dist), function(fc, point, dist){
+  map2(fc, fc_dist, function(fc, dist){
     dimnames(dist) <- dimnames(fc[[distribution_var(fc)]])
     fc[[distribution_var(fc)]] <- dist
-    point_fc <- compute_point_forecasts(dist, point_method)
+    point_fc <- compute_point_forecasts(dist, point_forecast)
     fc[names(point_fc)] <- point_fc
     fc
   })
 }
 
 build_smat_rows <- function(key_data){
+  lifecycle::deprecate_warn("0.2.1", "fabletools::build_smat_rows()", "fabletools::build_key_data_smat()")
   row_col <- sym(colnames(key_data)[length(key_data)])
   
   smat <- key_data %>%
@@ -299,12 +390,13 @@ build_key_data_smat <- function(x){
     agg_idx <- level[["loc"]][[1]]
     pos <- vec_match(x_leaf[disagg_col], x[agg_idx, disagg_col])
     pos <- vec_group_loc(pos)
+    pos <- pos[!is.na(pos$key),]
     # Add non-matches as leaf nodes
     agg_leaf <- setdiff(seq_along(agg_idx), pos$key)
     if(!is_empty(agg_leaf)){
       pos <- vec_rbind(
         pos,
-        structure(list(key = agg_leaf, loc = as.list(agg_idx[agg_leaf])), 
+        structure(list(key = agg_leaf, loc = as.list(seq_along(agg_leaf) + nrow(x_leaf))), 
                   class = "data.frame", row.names = agg_leaf)
       )
       x_leaf <<- vec_rbind(
@@ -318,7 +410,7 @@ build_key_data_smat <- function(x){
     abort("An error has occurred when constructing the summation matrix.\nPlease report this bug here: https://github.com/tidyverts/fabletools/issues")
   }
   idx_leaf <- vec_c(!!!x_leaf$.rows)
-  x$.rows[vec_c(!!!grp$loc)] <- vec_c(!!!grp$match)
+  x$.rows[unlist(x$.rows)[vec_c(!!!grp$loc)]] <- vec_c(!!!grp$match)
   return(list(agg = x$.rows, leaf = idx_leaf))
   # out <- matrix(0L, nrow = nrow(x), ncol = length(idx_leaf))
   # out[nrow(x)*(vec_c(!!!x$.rows)-1) + rep(seq_along(x$.rows), lengths(x$.rows))] <- 1L