+++ R/residuals.R
@@ -9,14 +9,12 @@
 #' @importFrom stats residuals
 #' @export
 residuals.mdl_df <- function(object, ...){
-  out <- gather(object, ".model", ".fit", !!!syms(mable_vars(object)))
-  kv <- key_vars(out)
-  out <- transmute(as_tibble(out),
-    !!!syms(kv),
-    !!sym(".model"),
-    residuals = map(!!sym(".fit"), residuals, ...)
-  )
-  unnest_tsbl(out, "residuals", parent_key = kv)
+  mbl_vars <- mable_vars(object)
+  kv <- key_vars(object)
+  object <- mutate(as_tibble(object), 
+              dplyr::across(all_of(mbl_vars), function(x) lapply(x, residuals, ...)))
+  object <- pivot_longer(object, mbl_vars, names_to = ".model", values_to = ".resid")
+  unnest_tsbl(object, ".resid", parent_key = c(kv, ".model"))
 }
 
 #' @param type The type of residuals to compute. If `type="response"`, residuals on the back-transformed data will be computed.
@@ -31,12 +29,10 @@ residuals.mdl_ts <- function(object, type = "innovation", ...){
   else{
     .resid <- residuals(object$fit, type = type, ...)
     if(is.null(.resid)){
-        warn(sprintf(
-'Residuals of type `%s` are not supported for %s models.
-Defaulting to `type="response"`', type, model_sum(object)))
-      .resid <- response(object)
-      .fits <- fitted(object)
-      .resid <- as.matrix(.resid[measured_vars(.resid)]) - as.matrix(.fits[measured_vars(.fits)])
+#         warn(sprintf(
+# 'Residuals of type `%s` are not supported for %s models.
+# Defaulting to `type="response"`', type, model_sum(object)))
+      return(residuals(object, type = "response", ...))
     }
   }
   .resid <- as.matrix(.resid)